+++ ca/django_ca/management/commands/resign_cert.py
@@ -74,6 +74,9 @@ default profile, currently {ca_settings.CA_DEFAULT_PROFILE}."""
         # Certificate Policies extension
         certificate_policies: Optional[x509.CertificatePolicies],
         certificate_policies_critical: bool,
+        # CRL Distribution Points extension
+        crl_full_names: Optional[List[x509.GeneralName]],
+        crl_distribution_points_critical: bool,
         # Extended Key Usage extension
         extended_key_usage: Optional[x509.ExtendedKeyUsage],
         extended_key_usage_critical: bool,
@@ -116,6 +119,13 @@ default profile, currently {ca_settings.CA_DEFAULT_PROFILE}."""
 
         if certificate_policies is not None:
             self._add_extension(extensions, certificate_policies, certificate_policies_critical)
+        if crl_full_names is not None:
+            distribution_point = x509.DistributionPoint(
+                full_name=crl_full_names, relative_name=None, crl_issuer=None, reasons=None
+            )
+            self._add_extension(
+                extensions, x509.CRLDistributionPoints([distribution_point]), crl_distribution_points_critical
+            )
         if extended_key_usage is not None:
             self._add_extension(extensions, extended_key_usage, extended_key_usage_critical)
         if issuer_alternative_name is not None: